# Rewrite GRLevelData::plus with OpenMP and SIMD