--- a/Source/GRChomboCore/GRLevelData.hpp
+++ b/Source/GRChomboCore/GRLevelData.hpp
@@ -20,8 +20,9 @@ class GRLevelData : public LevelData<FArrayBox>
 
     void setVal(const double a_val, const Interval a_comps);
 
-    // a_src and this must have the same box layout
-    void plus(const GRLevelData &a_src, const double a_scale);
+    // loop only goes over a_disjoint_box_layout
+    void plus(const GRLevelData &a_src, const double a_scale,
+              const DisjointBoxLayout &a_disjoint_box_layout);
 };
 
 #endif /* GRLEVELDATA_HPP_ */