Note that the prototype has also changed so that this function now only loops over the `Box`es in the passed `DisjointBoxLayout` rather than the intersection of the two `FArrayBox`es' `Box`es. This saves a little in not looping through the interior ghost cells but in `GRAMRLevel::updateODE`, `m_grown_grids` is passed which includes the outer boundary ghosts in the case of Sommerfeld BCs (and extrapolating/mixed BCs once #140 is merged). As far as I can see this should be the only case where we want to add the RHS to `soln` in the outer boundary ghosts.

It was probably unncessary but I generated an Intel compiler optimisation report to confirm the loop is being vectorized (although I suspect the old loop was probably being autovectorized anyway): [GRLevelData.optrpt](https://github.com/GRChombo/GRChombo/files/5719631/GRLevelData.optrpt.txt). Note that the lines numbers will differ a bit to what is in this PR as this was before I added the `#if CH_SPACEDIM == 3` macros.

I also had a look at the Chombo timer output for one of my own BBH examples over 8 timesteps with a few different numbers of OpenMP threads. Summing over all of the `GRAMRLevel::updateODE`s, the time taken in each case (on rank 0)

**Before changes**
1 thread: 59.09279s
2 threads: 50.31753s
4 threads: 49.14452s
6 threads: 48.26825s

**After changes**
1 thread: 54.51439s
2 threads: 28.94410s
4 threads: 18.28667s
6 threads: 15.18297s

@TaigoFr, I attempted to make this compatible with 2D but since you are more familiar with that than me, would you be able to check this is correct?