@mirenradia 2020-12-19
I should also add that I verified that I get the same results by comparing a t>0 checkpoint from before and after these changes with the HDF5 comparison tool. With the Intel compiler (and `-O3` enabled), I got agreement up to `5.0e-13`.