@TaigoFr 2020-12-20
Yes the compatibility with 2D seems good. After this is merged, when I add this to some code with a 2D example with evolution, I can also make an HDF5 comparison just to be really sure, but for now seems good!