@KAClough 2021-01-18
As discussed I am happy with the use of m_grown_grids, looks good to merge!